+++ packages/babel-preset-next-babel/test/integration.test.ts
@@ -1,9 +1,9 @@
 import { pluginTester } from 'babel-plugin-tester';
-import type { NextBabelPresetOptions } from 'pkgverse/next-babel/src/index';
+import type { NextBabelPresetOptions } from 'pkgverse/babel-preset-next-babel/src/index';
 
 // eslint-disable-next-line jest/require-hook
 pluginTester({
-  preset: require('@xunnamius/next-babel').default,
+  preset: require('@xunnamius/babel-preset-next-babel').default,
   presetName: 'next/babel',
   presetOptions: {} satisfies NextBabelPresetOptions,
   tests: {